Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/intrinsicsize/intrinsicsize-without-unsized-media.tentative.https.sub-expected.txt

  

FAIL Test intrinsicsize for html image element, src=/images/green.svg assert_equals: width expected 400 but got 100
FAIL Test intrinsicsize for html image element, src=/images/green.png assert_equals: width expected 400 but got 100
